requirements.txt
streamlit
openai
pdfplumber
langchain
chromadb
groq
python-dotenv
pytest

